# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Email Engine Core",
        "description": "payEase Payment Service.",
        "version": "1.0.0",
        "contact": {
            "name": "Email Engine Core",
            "email": "email@email.com",
            "url": "https://emailclientcore-2.onrender.com/"
        }
    },
    "servers": [
        {
            "url": "https://emailclientcore-2.onrender.com/"
          },
          {
            "url": "http://localhost:3000"
          }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer"
            },
            "noauthAuth": {
              "type": "http",
              "scheme": "noauth"
            }
          }
    },
    "tags": [
          {
            "name": "Users"
          }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [ "Users" ],
                "summary": "Users Registration",
                "description": "This endpoint is used to Register New Users in to the application.",
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "type": "object",
                          "example": {
                             "email":"",
                            "password": "",
                            "emailService":""
                          }
                        }
                      }
                    }
                  },     
                  "responses": {
                    "201": {
                      "description": "Created",
                      "headers": {
                        "Content-Security-Policy": {
                          "schema": {
                            "type": "string",
                            "example": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                          }
                        },
                        "X-DNS-Prefetch-Control": {
                          "schema": {
                            "type": "string",
                            "example": "off"
                          }
                        },
                        "Expect-CT": {
                          "schema": {
                            "type": "string",
                            "example": "max-age=0"
                          }
                        },
                        "X-Frame-Options": {
                          "schema": {
                            "type": "string",
                            "example": "SAMEORIGIN"
                          }
                        },
                        "Strict-Transport-Security": {
                          "schema": {
                            "type": "string",
                            "example": "max-age=15552000; includeSubDomains"
                          }
                        },
                        "X-Download-Options": {
                          "schema": {
                            "type": "string",
                            "example": "noopen"
                          }
                        },
                        "X-Content-Type-Options": {
                          "schema": {
                            "type": "string",
                            "example": "nosniff"
                          }
                        },
                        "X-Permitted-Cross-Domain-Policies": {
                          "schema": {
                            "type": "string",
                            "example": "none"
                          }
                        },
                        "Referrer-Policy": {
                          "schema": {
                            "type": "string",
                            "example": "no-referrer"
                          }
                        },
                        "X-XSS-Protection": {
                          "schema": {
                            "type": "integer",
                            "example": "0"
                          }
                        },
                        "Access-Control-Allow-Origin": {
                          "schema": {
                            "type": "string",
                            "example": "*"
                          }
                        },
                        "Content-Type": {
                          "schema": {
                            "type": "string",
                            "example": "application/json; charset=utf-8"
                          }
                        },
                        "Content-Length": {
                          "schema": {
                            "type": "integer",
                            "example": "24"
                          }
                        },
                        "ETag": {
                          "schema": {
                            "type": "string",
                            "example": "W/\"18-6mHpC91Ut4V5aiC0eWVy7rvTHFM\""
                          }
                        },
                        "Date": {
                          "schema": {
                            "type": "string",
                            "example": "Thu, 24 Nov 2022 08:39:00 GMT"
                          }
                        },
                        "Connection": {
                          "schema": {
                            "type": "string",
                            "example": "keep-alive"
                          }
                        },
                        "Keep-Alive": {
                          "schema": {
                            "type": "string",
                            "example": "timeout=5"
                          }
                        }
                      },
                      "content": {
                        "application/json": {
                          "schema": {
                            "type": "object"
                          },
                            "example": {
                                "message": "User registered",
                                "userId": "4b14d3ad-25fe-46b6-b52b-f060b6fa1d25"
                            }
                        }
                      }
                    }
                  }
            }
        },
        "/emails": {
            "get": {
              "tags": [
                "Users"
              ],
              "summary": "Get All Synced Emails",
              "requestBody": {
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": {
                        "email":"",
                        "password": "",
                        "emailService":"",
                        "userId":""
                      }
                    }
                  }
                }
              },  
              "responses": {
                "200": {
                  "description": "OK",
                  "headers": {
                    "Content-Security-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                      }
                    },
                    "X-DNS-Prefetch-Control": {
                      "schema": {
                        "type": "string",
                        "example": "off"
                      }
                    },
                    "Expect-CT": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=0"
                      }
                    },
                    "X-Frame-Options": {
                      "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                      }
                    },
                    "Strict-Transport-Security": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                      }
                    },
                    "X-Download-Options": {
                      "schema": {
                        "type": "string",
                        "example": "noopen"
                      }
                    },
                    "X-Content-Type-Options": {
                      "schema": {
                        "type": "string",
                        "example": "nosniff"
                      }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                      "schema": {
                        "type": "string",
                        "example": "none"
                      }
                    },
                    "Referrer-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "no-referrer"
                      }
                    },
                    "X-XSS-Protection": {
                      "schema": {
                        "type": "integer",
                        "example": "0"
                      }
                    },
                    "Access-Control-Allow-Origin": {
                      "schema": {
                        "type": "string",
                        "example": "*"
                      }
                    },
                    "Content-Type": {
                      "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                      }
                    },
                    "Content-Length": {
                      "schema": {
                        "type": "integer",
                        "example": "252"
                      }
                    },
                    "ETag": {
                      "schema": {
                        "type": "string",
                        "example": "W/\"fc-QlM6DWIVDN9b8KcyPkXIaHmdr0Q\""
                      }
                    },
                    "Date": {
                      "schema": {
                        "type": "string",
                        "example": "Fri, 02 Dec 2022 17:56:01 GMT"
                      }
                    },
                    "Connection": {
                      "schema": {
                        "type": "string",
                        "example": "keep-alive"
                      }
                    },
                    "Keep-Alive": {
                      "schema": {
                        "type": "string",
                        "example": "timeout=5"
                      }
                    }
                  },
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object"
                      },
                      "example": {
                        "emails": [
                            {
                                "subject": "TEST TEST TEST ",
                                "from": {
                                    "value": [
                                        {
                                            "address": "TEST@accounts.google.com",
                                            "name": "TEST  Team"
                                        }
                                    ],
                                    "html": "<span class=\"mp_address_group\"><span class=\"mp_address_name\">The Google Account Team</span> &lt;<a href=\"mailto:no-reply@accounts.google.com\" class=\"mp_address_email\">no-reply@accounts.google.com</a>&gt;</span>",
                                    "text": "\"The Google Account Team\" <no-reply@accounts.google.com>"
                                },
                                "date": "2024-03-19T19:49:23.000Z"
                            },
                            {
                                "subject": "TEST TEST TEST ",
                                "from": {
                                    "value": [
                                        {
                                            "address": "TEST@accounts.google.com",
                                            "name": "TEST  Team"
                                        }
                                    ],
                                    "html": "<span class=\"mp_address_group\"><span class=\"mp_address_name\">The Google Account Team</span> &lt;<a href=\"mailto:no-reply@accounts.google.com\" class=\"mp_address_email\">no-reply@accounts.google.com</a>&gt;</span>",
                                    "text": "\"The Google Account Team\" <no-reply@accounts.google.com>"
                                },
                                "date": "2024-03-19T19:49:23.000Z"
                            },
                            {
                                "subject": "TEST TEST TEST ",
                                "from": {
                                    "value": [
                                        {
                                            "address": "TEST@accounts.google.com",
                                            "name": "TEST  Team"
                                        }
                                    ],
                                    "html": "<span class=\"mp_address_group\"><span class=\"mp_address_name\">The Google Account Team</span> &lt;<a href=\"mailto:no-reply@accounts.google.com\" class=\"mp_address_email\">no-reply@accounts.google.com</a>&gt;</span>",
                                    "text": "\"The Google Account Team\" <no-reply@accounts.google.com>"
                                },
                                "date": "2024-03-19T19:49:23.000Z"
                            }
                        ]
                      }
                    }
                  }
                }
              }
            }
        },
        "/users": {
            "get": {
              "tags": [
                "Users"
              ],
              "summary": "Get All App Users",
              "requestBody": {
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": {
                        "indexName":"users"
                      }
                    }
                  }
                }
              },  
              "responses": {
                "200": {
                  "description": "OK",
                  "headers": {
                    "Content-Security-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                      }
                    },
                    "X-DNS-Prefetch-Control": {
                      "schema": {
                        "type": "string",
                        "example": "off"
                      }
                    },
                    "Expect-CT": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=0"
                      }
                    },
                    "X-Frame-Options": {
                      "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                      }
                    },
                    "Strict-Transport-Security": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                      }
                    },
                    "X-Download-Options": {
                      "schema": {
                        "type": "string",
                        "example": "noopen"
                      }
                    },
                    "X-Content-Type-Options": {
                      "schema": {
                        "type": "string",
                        "example": "nosniff"
                      }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                      "schema": {
                        "type": "string",
                        "example": "none"
                      }
                    },
                    "Referrer-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "no-referrer"
                      }
                    },
                    "X-XSS-Protection": {
                      "schema": {
                        "type": "integer",
                        "example": "0"
                      }
                    },
                    "Access-Control-Allow-Origin": {
                      "schema": {
                        "type": "string",
                        "example": "*"
                      }
                    },
                    "Content-Type": {
                      "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                      }
                    },
                    "Content-Length": {
                      "schema": {
                        "type": "integer",
                        "example": "252"
                      }
                    },
                    "ETag": {
                      "schema": {
                        "type": "string",
                        "example": "W/\"fc-QlM6DWIVDN9b8KcyPkXIaHmdr0Q\""
                      }
                    },
                    "Date": {
                      "schema": {
                        "type": "string",
                        "example": "Fri, 02 Dec 2022 17:56:01 GMT"
                      }
                    },
                    "Connection": {
                      "schema": {
                        "type": "string",
                        "example": "keep-alive"
                      }
                    },
                    "Keep-Alive": {
                      "schema": {
                        "type": "string",
                        "example": "timeout=5"
                      }
                    }
                  },
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object"
                      },
                      "example": {
                       "results": [
                            {
                                "_index": "users",
                                "_id": "d401f5f9-4b17-9ddf-b06bdbb4e8a8",
                                "_score": 1,
                                "_source": {
                                    "email": "test.org@gmail.com"
                                }
                            },
                            {
                                "_index": "users",
                                "_id": "4b14d3ad-25fe-b52b-f060b6fa1d25",
                                "_score": 1,
                                "_source": {
                                    "email": "test.org@gmail.com"
                                }
                            }
                        ]
                      }
                    }
                  }
                }
              }
            }
        },
        "/user": {
            "get": {
              "tags": [
                "Users"
              ],
              "summary": "Get An App Users",
              "requestBody": {
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": {
                        "indexName":"users",
                        "userId": ""
                      }
                    }
                  }
                }
              },  
              "responses": {
                "200": {
                  "description": "OK",
                  "headers": {
                    "Content-Security-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                      }
                    },
                    "X-DNS-Prefetch-Control": {
                      "schema": {
                        "type": "string",
                        "example": "off"
                      }
                    },
                    "Expect-CT": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=0"
                      }
                    },
                    "X-Frame-Options": {
                      "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                      }
                    },
                    "Strict-Transport-Security": {
                      "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                      }
                    },
                    "X-Download-Options": {
                      "schema": {
                        "type": "string",
                        "example": "noopen"
                      }
                    },
                    "X-Content-Type-Options": {
                      "schema": {
                        "type": "string",
                        "example": "nosniff"
                      }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                      "schema": {
                        "type": "string",
                        "example": "none"
                      }
                    },
                    "Referrer-Policy": {
                      "schema": {
                        "type": "string",
                        "example": "no-referrer"
                      }
                    },
                    "X-XSS-Protection": {
                      "schema": {
                        "type": "integer",
                        "example": "0"
                      }
                    },
                    "Access-Control-Allow-Origin": {
                      "schema": {
                        "type": "string",
                        "example": "*"
                      }
                    },
                    "Content-Type": {
                      "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                      }
                    },
                    "Content-Length": {
                      "schema": {
                        "type": "integer",
                        "example": "252"
                      }
                    },
                    "ETag": {
                      "schema": {
                        "type": "string",
                        "example": "W/\"fc-QlM6DWIVDN9b8KcyPkXIaHmdr0Q\""
                      }
                    },
                    "Date": {
                      "schema": {
                        "type": "string",
                        "example": "Fri, 02 Dec 2022 17:56:01 GMT"
                      }
                    },
                    "Connection": {
                      "schema": {
                        "type": "string",
                        "example": "keep-alive"
                      }
                    },
                    "Keep-Alive": {
                      "schema": {
                        "type": "string",
                        "example": "timeout=5"
                      }
                    }
                  },
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object"
                      },
                      "example": {
                       "results": 
                            {
                                "_index": "users",
                                "_id": "d401f5f9-4b17-9ddf-b06bdbb4e8a8",
                                "_score": 1,
                                "_source": {
                                    "email": "test.org@gmail.com"
                                }
                            }
                      }
                    }
                  }
                }
              }
            }
        }
    }

}